pyproject.toml
[tool.poetry]
name = "epic events"
version = "0.0.1"
description = "Internal CRM software from Epic Events."
authors = ["Théo Pidoux"]
readme = "README.md"
packages = [{include = "epic_events"}]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.20"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
click = "^8.1.7"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"
rich = "^13.6.0"
sentry-sdk = "^1.35.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
flake8 = "^6.1.0"
flake8-html = "^0.4.3"
coverage = "^7.3.1"
pytest-sqlalchemy-mock = "^0.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
